create tools and lib folder and move relevant files
