+++ src/tools/python/extract_tooltips.py
